+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file summarises the most important changes that went live on each release
 of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
+## Unreleased
+### Fixed
+- Proper order of table columns for sync_process/index page
+
 ## [0.11.5] - 2016-09-28
 ### Changed
 - Workers marked as `queued` to avoid duplicated messages on SQS.